core/requirements.txt
asgiref==3.8.1
cffi==1.17.1
coverage==7.9.1
cryptography==45.0.4
dj-database-url==3.0.0
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
Markdown==3.8.1
packaging==25.0
pillow==11.2.1
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0